Refactor app to separate file
